# Faster reverse_dict