- [x] Tests added / passed
- [x] Passes `black dask` / `flake8 dask`

Use a dictionary view union to avoid the creation of three lists.  An additional optimization caches the set.add method as a local variable.

Benchmarks:
```
# S = small graph of size 10
# B = big graph of size 100
# VB = very big graph of size 1000
# All graphs randomly generated by networkx.scale_free_graph.  Tested with Python 3.7
In [273]: %timeit reverse_dict(S)
6.24 µs ± 319 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [274]: %timeit faster_reverse_dict(S)
4.63 µs ± 41.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [275]: %timeit reverse_dict(B)
55.7 µs ± 440 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [276]: %timeit faster_reverse_dict(B)
41.8 µs ± 1.82 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [277]: %timeit reverse_dict(VB)
660 µs ± 19.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [278]: %timeit faster_reverse_dict(VB)
471 µs ± 23.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [279]: %timeit cached_reverse_dict(S)
4.43 µs ± 34 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [280]: %timeit cached_reverse_dict(B)
38 µs ± 286 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [281]: %timeit cached_reverse_dict(VB)
448 µs ± 14.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```