@TomAugspurger 2019-10-08
Looks like this caused a doctest failure: https://travis-ci.org/dask/dask/jobs/595217750#L1191, as the order of the elements is now different. Presumably you can just skip the doctest (or sort the output by key).

cc @jcrist if you have a chance to glance through the changes.