@groutr 2019-10-08
I think this can be made faster with a defaultdict.