@groutr 2019-10-22
@TomAugspurger is there anything further that I need to do?  I'd be happy to change back to a regular dictionary if a defaultdict is not desired.  The defaultdict does provide slightly better performance and with no default_factory, it behaves like a regular dictionary.