@groutr 2019-10-22
In summary, the main benefit of this PR is that we only iterate over the graph once when reversing the edges.  This PR also avoids the unnecessary list allocations from the original version.
For small dictionaries, ~26% performance gain and for larger dictionaries ~32% performance gain.