@TomAugspurger 2019-10-22
This seems reasonable. cc @mrocklin for on an opinion on whether `reverse_dict` preserving the order of the keys is valuable or not.