@mrocklin 2019-10-22
> cc @mrocklin for on an opinion on whether reverse_dict preserving the order of the keys is valuable or not.

I don't think so?